moved flex wrappers to expect actual, testing on jni now. should be able to write decoder post that

+++ flatinvoker-core/cpp/common/Flex.cpp
@@ -1,17 +1,8 @@
 #include <common/Flex.h>
-#include <common/CppBase.h>
+#include <common/CppBase.hpp>
 #include <flatbuffers/flexbuffers.h>
 #include <flatbuffers/idl.h>
 #include <flatbuffers/buffer.h>
-//#include <jsi/jsi.h>
-//
-//void bindJSIModule() {
-//    auto undefined = facebook::jsi::Value();
-//    auto null = facebook::jsi::Value(nullptr);
-//    if (undefined.isUndefined() == null.isNull()) {
-//        printf("Random code to check if jsi is linked");
-//    }
-//}
 
 /*
  * Optimise all these, inline what you can.
@@ -22,43 +13,43 @@ struct BuilderInfo {
 
 class FlexBufferStore {
     // Merge ?
-    std::unordered_map<long, std::shared_ptr<flexbuffers::Builder>> builders;
-    std::unordered_map<long, BuilderInfo> builderInfo;
+    std::unordered_map<uintptr_t, std::shared_ptr<flexbuffers::Builder>> builders;
+    std::unordered_map<uintptr_t, BuilderInfo> builderInfo;
 
 public:
-    long Create() {
+    uintptr_t Create() {
         auto builder = std::make_shared<flexbuffers::Builder>(1024);
-        auto pointer = reinterpret_cast<long>(builder.get());
+        auto pointer = reinterpret_cast<uintptr_t>(builder.get());
         builders[pointer] = builder;
         builderInfo[pointer] = {.isFinished = false};
         return pointer;
     }
 
     // Also need to destroy when builder goes out of scope in parent language
-    void Destroy(long pointer) {
+    void Destroy(uintptr_t pointer) {
         auto builder = Get(pointer);
         builder->Clear();
         builders.erase(pointer);
         builderInfo.erase(pointer);
     }
 
-    shared_ptr<flexbuffers::Builder> Get(long pointer) {
+    shared_ptr<flexbuffers::Builder> Get(uintptr_t pointer) {
         if (!builders.contains(pointer)) return nullptr;
         return builders[pointer];
     }
 
-    inline shared_ptr<flexbuffers::Builder> GetOrThrow(long pointer) {
+    inline shared_ptr<flexbuffers::Builder> GetOrThrow(uintptr_t pointer) {
         auto builder = Get(pointer);
         GUARD_THROW(builder, "No FlexBuilder found");
         return builder;
     }
 
-    bool IsFinished(long pointer) {
+    bool IsFinished(uintptr_t pointer) {
         if (!builderInfo.contains(pointer)) return false;
         return builderInfo[pointer].isFinished;
     }
 
-    void Finish(long pointer) {
+    void Finish(uintptr_t pointer) {
         auto builder = GetOrThrow(pointer);
         if (IsFinished(pointer)) return;
         // log double finish error
@@ -69,7 +60,7 @@ public:
 
 } FlexStore;
 
-void Flex_ParseJson(long pointer, const char *jsonString) {
+void Flex_ParseJson(uintptr_t pointer, const char *jsonString) {
     auto builder = FlexStore.Get(pointer);
     flatbuffers::Parser parser;
     parser.ParseFlexBuffer(jsonString, nullptr, builder.get());
@@ -77,24 +68,24 @@ void Flex_ParseJson(long pointer, const char *jsonString) {
     Flex_Finish(pointer);
 }
 
-long Flex_Create() {
+uintptr_t Flex_Create() {
     return FlexStore.Create();
 }
 
-void Flex_Destroy(long pointer) {
+void Flex_Destroy(uintptr_t pointer) {
     FlexStore.Destroy(pointer);
 }
 
-void Flex_Finish(long pointer) {
+void Flex_Finish(uintptr_t pointer) {
     FlexStore.Finish(pointer);
 }
 
-FlexArray Flex_GetBuffer(long pointer) {
+FlexArray Flex_GetBuffer(uintptr_t pointer) {
     auto builder = FlexStore.GetOrThrow(pointer);
     return { builder->GetBuffer().data(), builder->GetSize() };
 }
 
-void Flex_Null(long pointer, const char* key) {
+void Flex_Null(uintptr_t pointer, const char* key) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Null();
@@ -103,7 +94,7 @@ void Flex_Null(long pointer, const char* key) {
     }
 }
 
-void Flex_Int(long pointer, const char* key, int64_t value) {
+void Flex_Int(uintptr_t pointer, const char* key, int64_t value) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Int(value);
@@ -112,7 +103,7 @@ void Flex_Int(long pointer, const char* key, int64_t value) {
     }
 }
 
-void Flex_Float(long pointer, const char* key, float value) {
+void Flex_Float(uintptr_t pointer, const char* key, float value) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Float(value);
@@ -121,7 +112,7 @@ void Flex_Float(long pointer, const char* key, float value) {
     }
 }
 
-void Flex_Double(long pointer, const char* key, double value) {
+void Flex_Double(uintptr_t pointer, const char* key, double value) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Double(value);
@@ -130,7 +121,7 @@ void Flex_Double(long pointer, const char* key, double value) {
     }
 }
 
-void Flex_Bool(long pointer, const char* key, bool value) {
+void Flex_Bool(uintptr_t pointer, const char* key, bool value) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Bool(value);
@@ -139,7 +130,7 @@ void Flex_Bool(long pointer, const char* key, bool value) {
     }
 }
 
-void Flex_String(long pointer, const char* key, const char* value) {
+void Flex_String(uintptr_t pointer, const char* key, const char* value) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->String(value);
@@ -148,7 +139,7 @@ void Flex_String(long pointer, const char* key, const char* value) {
     }
 }
 
-void Flex_Blob(long pointer, const char* key, const uint8_t* value, size_t length) {
+void Flex_Blob(uintptr_t pointer, const char* key, const uint8_t* value, size_t length) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         builder->Blob(value, length);
@@ -157,7 +148,7 @@ void Flex_Blob(long pointer, const char* key, const uint8_t* value, size_t lengt
     }
 }
 
-size_t Flex_StartMap(long pointer, const char* key) {
+size_t Flex_StartMap(uintptr_t pointer, const char* key) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         return builder->StartMap();
@@ -166,12 +157,12 @@ size_t Flex_StartMap(long pointer, const char* key) {
     }
 }
 
-void Flex_EndMap(long pointer, size_t mapStart) {
+void Flex_EndMap(uintptr_t pointer, size_t mapStart) {
     auto builder = FlexStore.GetOrThrow(pointer);
     builder->EndMap(mapStart);
 }
 
-size_t Flex_StartVector(long pointer, const char* key) {
+size_t Flex_StartVector(uintptr_t pointer, const char* key) {
     auto builder = FlexStore.GetOrThrow(pointer);
     if (key == nullptr) {
         return builder->StartVector();
@@ -180,7 +171,7 @@ size_t Flex_StartVector(long pointer, const char* key) {
     }
 }
 
-void Flex_EndVector(long pointer, size_t vectorStart) {
+void Flex_EndVector(uintptr_t pointer, size_t vectorStart) {
     auto builder = FlexStore.GetOrThrow(pointer);
     builder->EndVector(vectorStart, false, false);
 }
